Add routing to new Artwork
