+++ src/Styleguide/Pages/Artwork/Components/__stories__/ArtistInfo.story.tsx
@@ -13,7 +13,7 @@ storiesOf("Styleguide/Artwork", module).add("ArtistInfo", () => {
       <Section title="Artist Info">
         <Grid>
           <Row>
-            <Col xl={8} lg={8} md={8} sm={12} xs={12}>
+            <Col md={8}>
               <ArtistInfo
                 name="Francesca DiMattio"
                 insights={insights}
